backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "httpayer"
version = "0.1.0"
description = "Python Library for HttPayer Server"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"

authors = [
    { name = "Brandyn Hamilton", email = "brandynham1120@gmail.com"}
]

dependencies = [
    "pandas",
    "numpy",
    "python-dotenv",
    "flask[async]",
    "web3",
    "diskcache>=5.6.3",
    "click",
    "ccip-terminal==0.1.3",
    "apscheduler>=3.10.4",
    "cachetools>=5.5.2",
    "pyngrok>=7.0.5",
    "python-viem>=0.1.0",
    "pywin32; sys_platform == 'win32'",
    "chartengineer==0.1.3",
]

[tool.setuptools.packages.find]
include = ["httpayer", "httpayer.*"]
